+++ src/main.tsx
@@ -1,7 +1,7 @@
 // import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import { TestsUI } from './App';
+import { TestsUI } from './App/TestsUI';
 import { StrictMode } from 'react';
 
 createRoot(document.getElementById('root')!).render(
